requirements.txt
requests
web3
python-dotenv
schedule
